+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2023-05-11
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): Command system.
+
+### Changed
+- [@mjkuranda](https://github.com/mjkuranda): More significant role for a `RetroGraphicsManager` class.
+
+### Removed:
+- [@mjkuranda](https://github.com/mjkuranda): Static generating palettes and converting images.
+
 ## [2.3.0] - 2023-05-10
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Factory for creating palettes.